fix(array): change size_t to uint32_t
in order to produce correct results with BH_PARENT decrease
possible size of array from size_t to uint32_t and use int64_t
as an index passed to bh_sift_down/up - (i - 1) / 2 should give
0 in case of i equals 0)

+++ includes/array.h
@@ -9,8 +9,8 @@
 typedef struct	s_array
 {
 	void		**data;
-	size_t		len;
-	size_t		cap;
+	uint32_t	len;
+	uint32_t	cap;
 	void		(*del)(void*);
 	int			(*cmp)(void*, void*);
 }				t_array;
